Develop

+++ src/Listeners/AddressCreateOrUpdateDeliveryListener.php
@@ -18,26 +18,32 @@ class AddressCreateOrUpdateDeliveryListener
             $created = !is_null($event->model->address);
             $chargeAddresses = $event->request->input('address_delivery', []);
 
-            if(!is_null($event->model->getOriginal('deleted_at'))){
+            if (!is_null($event->model->getOriginal('deleted_at'))) {
                 return;
             }
 
+            if(count($chargeAddresses) > 0){
 
-            if ($created) {
-                $event->model->addressDelivery()->delete();
+                if ($created) {
+                    $event->model->addressDelivery()->delete();
+                }
+
+                foreach ($chargeAddresses as $address) {
+                    $address = \RiseTechApps\Address\Address::fillWithDefault($address, $event->model);
+
+                    $address['address_type'] = get_class($event->model);
+                    $address['address_id'] = $event->model->getKey();
+                    $address['type'] = 'delivery';
+                    AddressModel::create($address);
+                }
             }
-
-            foreach ($chargeAddresses as $address) {
-                $address = \RiseTechApps\Address\Address::fillWithDefault($address, $event->model);
-
-                $address['address_type'] = get_class($event->model);
-                $address['address_id'] = $event->model->getKey();
-                $address['type'] = 'delivery';
-                AddressModel::create($address);
-            }
-
 
         } catch (\Exception $exception) {
+            logglyError()->exception($exception)
+                ->withRequest($event->request)
+                ->performedOn(static::class)
+                ->log("Error registering address");
+
         }
     }
 }
